modularize core style font_pair
