+++ db/migrate/20220718211646_rename_secondary_back_up_into_secondary_backup_in_font_pairs.rb
@@ -1,0 +1,5 @@
+class RenameSecondaryBackUpIntoSecondaryBackupInFontPairs < ActiveRecord::Migration[7.0]
+  def change
+    rename_column :font_pairs, :secondary_back_up, :secondary_backup
+  end
+end
