+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_07_18_205009) do
+ActiveRecord::Schema.define(version: 2022_07_18_211646) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -94,7 +94,7 @@ ActiveRecord::Schema.define(version: 2022_07_18_205009) do
     t.string "primary"
     t.string "primary_backup"
     t.string "secondary"
-    t.string "secondary_back_up"
+    t.string "secondary_backup"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
