+++ spec/models/core/pages/context_spec.rb
@@ -1,5 +1,23 @@
 # frozen_string_literal: true
 
+# == Schema Information
+#
+# Table name: contexts
+#
+#  id         :bigint           not null, primary key
+#  query      :text             not null
+#  created_at :datetime         not null
+#  updated_at :datetime         not null
+#  page_id    :bigint           not null
+#
+# Indexes
+#
+#  index_contexts_on_page_id  (page_id)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (page_id => pages.id)
+#
 require 'rails_helper'
 require 'support/shared_example/models/should_have_database_colums'
 require 'support/shared_example/models/should_have_active_record_validations'
